app/src/main/java/com/example/gaijinsmash/transitapp/utils/EmailValidator.java
package com.example.gaijinsmash.transitapp.utils;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

/**
 * Created by ryanj on 8/10/2017.
 */

public class EmailValidator {
    private Pattern pattern;
    private Matcher matcher;
    private static final String EMAIL_PATTERN = "^[_A-Za-z0-9-]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";

    public boolean isEmailValid(String email) {
        pattern = Pattern.compile(EMAIL_PATTERN);
        matcher = pattern.matcher(email);

        return matcher.matches();
    }
}
